refactor: split oversized classes to satisfy SonarQube S1448 (max 20 methods)
- Split ExtractController (31 methods) into ExtractController (extraction
  pipeline), ApiExtractController (public token-authenticated API),
  ReferenceEditController (viewref page, save/autosave, BibTeX import, DOI
  enrich) and SemanticScholarImportController, sharing document-authorization
  helpers via the new DocumentAccessTrait. Route names/paths are unchanged.
- Split OpenAlexResolver (21 methods) into OpenAlexResolver (HTTP
  orchestration), OpenAlexWorkParser (pure payload parsing) and
  OpenAlexQuotaTracker (daily bulk-quota cache tracking); update
  config/services.yaml wiring and OpenAlexResolverTest accordingly.
